Remove unneeded delay in hspiread32
This reduces the time for a typical readCelcius call from over a
millisecond to a couple of dozen microseconds, making the library much
more suited for use in timing-critical applications and control loops.

This delay was present between the lowering of the chip select line and
the first transfer. Looking at the MAX31855 datashee, it shows that the
time needed between the chip select drop and the first clock rise (tCSS)
is 100ns. This delay will be present due to function call overheads many
times over, so there is no need for an explicit delay.

Similar delays exist for software SPI, which are probably also not
needed, but I did not test this.