Simplify CJT compensation
Simplified CJT linearization as described in the post:
https://forums.adafruit.com/viewtopic.php?f=19&t=32086&start=75#p579783
"Re: Max 31855 Reading High Temperatures

by zike on Sun Apr 23, 2017 7:24 pm
Many thanks for this! Wonderfully executed and exquisitely documented. I am happily using this routine with the MAX31855 breakout and a datalogging shield on an UNO to read, log, and safety-interlock my 1250C heat treating furnace.

I thought it may be helpful to point out that the first part of the code, inverting and linearizing the cold junction comp, is not strictly necessary. The fixed Seebeck coefficient used in the 31855 for CJC (40.73 uV/C) is chosen to afford less than 0.1C integral linearity error _over the service temperature range of the chip_ . By comparison, the ADC resolution corresponds to 0.25C, and the internal thermometer's absolute error can be ±2C to start with.

We therefore suffer no loss of accuracy by substituting the following line:
internalVoltage = internalTemp * 0.04073 ; // (mV) MAX38155 comp scale is 40.73 uV/C
or the entirety of Step 3. This saves considerable memory and execution time!

(The pseudo-floating point exp() and pow() are huge time sucks for a mere UNO).

I hope this can help your code fit into other apps (like mine) where RAM and loop time are on the bare edge.

Thanks again for all your hard work,

Mike