Allow code to specify which SPI buss to use
As mentioned in issue #43,

A PJRC Forum member ran into an issue on using this library on a Teensy 4.1
He was trying to use SPI1 for this, but there was no way to tell the underlying spi_dev object to use SPI1, so I added an additional constructor.

Not sure if the SPI device would be best as first or last paremeter (I choose last)
and if better to pass by pointer or by reference.

It compiled for a Teensy, but have no hardware to test it with.